rename folder to pdr dashboard
