+++ pdr_backend/pdr_dashboard/dash_components/view_elements.py
