+++ pdr_backend/pdr_dashboard/test/test_util.py
@@ -1,14 +1,14 @@
 from unittest.mock import patch
 
 from enforce_typing import enforce_types
-from pdr_backend.predictoor_dashboard.dash_components.util import (
+from pdr_backend.pdr_dashboard.dash_components.util import (
     _query_db,
     get_feeds_data_from_db,
     get_predictoors_data_from_db,
     get_payouts_from_db,
 )
 
-from pdr_backend.predictoor_dashboard.test.resources import (
+from pdr_backend.pdr_dashboard.test.resources import (
     _prepare_test_db,
     _clear_test_db,
 )
@@ -17,9 +17,7 @@ from pdr_backend.lake.payout import Payout
 
 
 @enforce_types
-@patch(
-    "pdr_backend.analytics.predictoor_dashboard.dash_components.util.DuckDBDataStore"
-)
+@patch("pdr_backend.pdr_dashboard.dash_components.util.DuckDBDataStore")
 def test_query_db(
     mock_duckdb_data_store,
 ):
